+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './init';
+export * from './main';
